refactor(build): remove extraneous command in RUN
`RUN` directive on L49 had remenants of a build that prepared for
`matplotlib` use with the image. This required a `.config/matplotlib`
folder with the correct permissions, and so was created during the
build. However, I do not want to make assumptions as to the use of this
image, and removed this from the build. Does not fix or improve anything...
just for organization's sake.